requirements.txt
flask==3.1.0
requests==2.32.3
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
